requirements.txt
tkinter
random
Pillow
os